Fix failing tests  - adapted to new return type

+++ cdmlib-services/src/test/java/eu/etaxonomy/cdm/api/service/IdentifiableServiceBaseTest.java
@@ -5,7 +5,7 @@
 *
 * The contents of this file are subject to the Mozilla Public License Version 1.1
 * See LICENSE.TXT at the top of this package for the full license terms.
-*/ 
+*/
 
 package eu.etaxonomy.cdm.api.service;
 
@@ -48,54 +48,54 @@ import eu.etaxonomy.cdm.test.integration.CdmTransactionalIntegrationTest;
 public class IdentifiableServiceBaseTest extends CdmTransactionalIntegrationTest {
 	@SuppressWarnings("unused")
 	private static final Logger logger = Logger.getLogger(IdentifiableServiceBaseTest.class);
-	
-	
+
+
 	@SpringBeanByType
 	private INameService nameService;
-	
+
 	@SpringBeanByType
 	private ITermService termService;
-	
+
 	@SpringBeanByType
 	private IVocabularyService vocService;
-	
+
 	@SpringBeanByType
 	private ITaxonService taxonService;
-	
+
 	@SpringBeanByType
 	private IClassificationService classificationService;
-	
+
 /****************** TESTS *****************************/
-	
+
 	@Test
 	public final void voidTestSeriveExists(){
 		Assert.assertNotNull("Service shoulb be initialized", nameService);
 	}
 
-	
+
 	@Test
 	@DataSet
-	@ExpectedDataSet	
+	@ExpectedDataSet
 	public final void testUpdateTitleCache() {
 		Assert.assertEquals("There should be 5 TaxonNames in the data set", 5, nameService.count(TaxonNameBase.class));
 		Class clazz = TaxonNameBase.class;
 		int stepSize = 2;
 		nameService.updateTitleCache(clazz, stepSize, null, null);
 		commit();
-//		commitAndStartNewTransaction(new String[]{"TaxonNameBase","TaxonNameBase_AUD"});	
+//		commitAndStartNewTransaction(new String[]{"TaxonNameBase","TaxonNameBase_AUD"});
 	}
-	
-	
+
+
 	@Test
 	@DataSet(value="IdentifiableServiceBaseTest.testFindByIdentifier.xml")
 	public final void testListByIdentifier(){
 		UUID uuidIdentifierType1 = UUID.fromString("02bb62db-a229-4eeb-83e6-a9a093943d5e");
 		UUID uuidIdentifierType2 = UUID.fromString("ef6e960f-5289-456c-b25c-cff7f4de2f63");
-		
-		
+
+
 		DefinedTerm it1 = (DefinedTerm)termService.find(uuidIdentifierType1);
 		Assert.assertNotNull("identifier type must not be null", it1);
-		
+
 		boolean includeEntity = true;
 		Pager<FindByIdentifierDTO<Taxon>> taxonPager = taxonService.findByIdentifier(Taxon.class, "ext-1234", it1, null, includeEntity, null, null, null);
 		Assert.assertTrue("Result should not be empty", taxonPager.getCount() == 1);
@@ -107,25 +107,25 @@ public class IdentifiableServiceBaseTest extends CdmTransactionalIntegrationTest
 		Identifier<?> identifier = taxon.getIdentifiers().get(0);
 		DefinedTerm type = CdmBase.deproxy(identifier.getType(), DefinedTerm.class);
 		Assert.assertEquals(uuidIdentifierType1, type.getUuid());
-		
+
 		Pager<FindByIdentifierDTO<TaxonNameBase>> names = nameService.findByIdentifier(
 				TaxonNameBase.class, "ext-1234", null, null, includeEntity, null, null, null);
 		Assert.assertTrue("Identifier does not exist for TaxonName", names.getCount() == 0);
-		
+
 		taxonPager = taxonService.findByIdentifier(null, "ext-1234", null, null, includeEntity, null, null, null);
 		Assert.assertEquals("Result size for 'ext-1234' should be 1", 1, taxonPager.getRecords().size());
-		
+
 		taxonPager = taxonService.findByIdentifier(Taxon.class, null, null, null, includeEntity, null, null, null);
 		Assert.assertEquals("Result should not be empty", 2 , taxonPager.getRecords().size());
-		
+
 		//includeEntity
 		includeEntity = false;
 		taxonPager = taxonService.findByIdentifier(Taxon.class, "ext-1234", it1, null, includeEntity, null, null, null);
 		entity = taxonPager.getRecords().get(0).getCdmEntity();
 		Assert.assertNull("Taxon must not be returned with includeEntity = false", entity.getEntity());
-		
-		
-		
+
+
+
 		//Matchmode
 		includeEntity = false;
 		MatchMode matchmode = null;
@@ -148,15 +148,15 @@ public class IdentifiableServiceBaseTest extends CdmTransactionalIntegrationTest
 		taxonPager = taxonService.findByIdentifier(null, "ext", null, MatchMode.BEGINNING, includeEntity, null, null, null);
 		Assert.assertEquals("Total result size for starts with 'ext' should be 4", 4, taxonPager.getRecords().size());
 		taxonPager = taxonService.findByIdentifier(null, "ext", null, MatchMode.BEGINNING, includeEntity, 2, 1, null);
-		Assert.assertEquals("Total result size for starts with 'ext' should be 4", Integer.valueOf(4), taxonPager.getCount());
+		Assert.assertEquals("Total result size for starts with 'ext' should be 4", Long.valueOf(4), taxonPager.getCount());
 		Assert.assertEquals("Result size for starts with 'ext' second page should be 2", Integer.valueOf(2), taxonPager.getPageSize());
 		Assert.assertEquals("The third taxon (first on second page) should be ext-syn1", "ext-syn1", taxonPager.getRecords().get(0).getIdentifier().getIdentifier());
-		
+
 		taxonPager = taxonService.findByIdentifier(Taxon.class, "ext", null, MatchMode.BEGINNING, includeEntity, null, null, null);
 		Assert.assertTrue("Result size for 'ext' should be 2", taxonPager.getCount() == 2);
 
 	}
-	
+
 	@Test
 	@DataSet(value="IdentifiableServiceBaseTest.testFindByIdentifier.xml")
 	public final void testListByIdentifierClassification(){
@@ -164,31 +164,31 @@ public class IdentifiableServiceBaseTest extends CdmTransactionalIntegrationTest
 		Classification classification = classificationService.find(5000);
 		TaxonNode rootNode = classification.getRootNode();
 		Pager<FindByIdentifierDTO<Taxon>> taxonPager = taxonService.findByIdentifier(Taxon.class, "ext-1234", null, rootNode, MatchMode.EXACT, false, null, null, null);
-		Assert.assertEquals("Result size for 'ext' should be 1", Integer.valueOf(1), taxonPager.getCount());
-		Assert.assertEquals("Result size for 'ext' should be 1", Integer.valueOf(1), (Integer)taxonPager.getRecords().size());
-		
+		Assert.assertEquals("Result size for 'ext' should be 1", Long.valueOf(1), taxonPager.getCount());
+		Assert.assertEquals("Result size for 'ext' should be 1", 1, taxonPager.getRecords().size());
+
 		Pager<FindByIdentifierDTO<Taxon>> taxPager = taxonService.findByIdentifier(Taxon.class, "ext-cache1", null, rootNode, MatchMode.EXACT, false, null, null, null);
-		Assert.assertEquals("Result size for 'ext' should be 0", Integer.valueOf(0), taxPager.getCount());
+		Assert.assertEquals("Result size for 'ext' should be 0", Long.valueOf(0), taxPager.getCount());
 		Assert.assertEquals("Result size for 'ext' should be 0", 0, taxPager.getRecords().size());
-		
+
 		rootNode = null;  //check against missing filter
 		taxPager = taxonService.findByIdentifier(Taxon.class, "ext-cache1", null, rootNode, MatchMode.EXACT, false, null, null, null);
-		Assert.assertEquals("Result size for 'ext-cache1' without filter should be 1", Integer.valueOf(1), taxPager.getCount());
+		Assert.assertEquals("Result size for 'ext-cache1' without filter should be 1", Long.valueOf(1), taxPager.getCount());
 		Assert.assertEquals("Result size for 'ext-cache1' without filter should be 1", 1, taxPager.getRecords().size());
-		
+
 		//TaxonBase
 		rootNode = classification.getRootNode();
 		Pager<FindByIdentifierDTO<TaxonBase>> tbPager = taxonService.findByIdentifier(TaxonBase.class, "ext-1234", null, rootNode, MatchMode.EXACT, false, null, null, null);
-		Assert.assertEquals("Result size for 'ext' should be 1", Integer.valueOf(1), tbPager.getCount());
-		Assert.assertEquals("Result size for 'ext' should be 1", Integer.valueOf(1), (Integer)tbPager.getRecords().size());
+		Assert.assertEquals("Result size for 'ext' should be 1", Long.valueOf(1), tbPager.getCount());
+		Assert.assertEquals("Result size for 'ext' should be 1", 1, tbPager.getRecords().size());
 
 		tbPager = taxonService.findByIdentifier(TaxonBase.class, "ext-cache1", null, rootNode, MatchMode.EXACT, false, null, null, null);
-		Assert.assertEquals("Result size for 'ext' should be 0", Integer.valueOf(0), tbPager.getCount());
+		Assert.assertEquals("Result size for 'ext' should be 0", Long.valueOf(0), tbPager.getCount());
 		Assert.assertEquals("Result size for 'ext' should be 0", 0, tbPager.getRecords().size());
-		
+
 		//Synonym
 		Pager<FindByIdentifierDTO<Synonym>> synPager = taxonService.findByIdentifier(Synonym.class, "ext-syn", null, rootNode, MatchMode.BEGINNING, false, null, null, null);
-		Assert.assertEquals("1 Synonym should be linked to the according classification", Integer.valueOf(1), synPager.getCount());
+		Assert.assertEquals("1 Synonym should be linked to the according classification", Long.valueOf(1), synPager.getCount());
 		Assert.assertEquals("1 Synonym should be linked to the according classification", 1, synPager.getRecords().size());
 
 	}
@@ -202,35 +202,35 @@ public class IdentifiableServiceBaseTest extends CdmTransactionalIntegrationTest
     @Override
     public void createTestDataSet() throws FileNotFoundException {
 		TermVocabulary<DefinedTerm> voc = vocService.find(VocabularyEnum.IdentifierType.getUuid());
-    	
+
 		DefinedTerm identifierType1 = DefinedTerm.NewIdentifierTypeInstance(null, "identifierType1", null);
     	voc.addTerm(identifierType1);
 		termService.save(identifierType1);
     	DefinedTerm identifierType2 = DefinedTerm.NewIdentifierTypeInstance(null, "identifierType2", null);
     	voc.addTerm(identifierType2);
 		termService.save(identifierType2);
-    	
-    	
+
+
     	BotanicalName name = BotanicalName.NewInstance(Rank.SPECIES());
         Taxon tb = Taxon.NewInstance(name, null);
         Identifier.NewInstance(tb, "ext-1234", identifierType1);
         Identifier.NewInstance(name, "ext-name12", identifierType2);
         taxonService.saveOrUpdate(tb);
-        
+
         Taxon tb2 = Taxon.NewInstance(null, null);
         tb2.setTitleCache("Cached taxon", true);
         Identifier.NewInstance(tb2, "ext-cache1", identifierType2);
         taxonService.saveOrUpdate(tb2);
-        
+
         Classification classification = Classification.NewInstance("My classification");
         classification.addChildTaxon(tb, null, null);
         classificationService.saveOrUpdate(classification);
-        
+
         tb2.addSynonymName(null, SynonymRelationshipType.HOMOTYPIC_SYNONYM_OF());
-        
+
         commitAndStartNewTransaction(null);
 
-        // this will write flat xml file to the same package in the test resources 
+        // this will write flat xml file to the same package in the test resources
         // the test file is named after the test class like: TestClassName.xml
 		writeDbUnitDataSetFile(new String[] {
 		        "TAXONBASE", "TAXONNAMEBASE","IDENTIFIER","TAXONBASE_IDENTIFIER",
@@ -241,7 +241,7 @@ public class IdentifiableServiceBaseTest extends CdmTransactionalIntegrationTest
 		        "TERMVOCABULARY",
 		        "SYNONYMRELATIONSHIP"
 		 }, "xxxx");
-            
+
     }
